Separate codegen module; to reduce 'cfg(not(test), ...)'
